repackage the implementation with static methods
