+++ lib/password_compat.php
@@ -1,16 +1,7 @@
 <?php
 
-if (version_compare(PHP_VERSION, '5.3.7', '<')) {
-	trigger_error("The Password Compatibility Library requires PHP >= 5.3.7", E_USER_WARNING);
-	// Prevent defining the functions
-	return;
-}
-
-defined('PASSWORD_BCRYPT') or define('PASSWORD_BCRYPT', 1);
-
-defined('PASSWORD_DEFAULT') or define('PASSWORD_DEFAULT', PASSWORD_BCRYPT);
-
-if (!function_exists('password_hash')) {
+class password_compat
+{
 	/**
 	 * Hash the password using the specified algorithm
 	 *
@@ -20,7 +11,7 @@ if (!function_exists('password_hash')) {
 	 *
 	 * @returns string|false The hashed password, or false on error.
 	 */
-	function password_hash($password, $algo, array $options = array()) {
+	static function hash($password, $algo, array $options = array()) {
 		if (!function_exists('crypt')) {
 			trigger_error("Crypt must be loaded for password_hash to function", E_USER_WARNING);
 			return null;
@@ -78,7 +69,7 @@ if (!function_exists('password_hash')) {
 				$salt = str_replace('+', '.', base64_encode($salt));
 			}
 		} else {
-			$salt = __password_make_salt($required_salt_len);
+			$salt = self::make_salt($required_salt_len);
 		}
 		$salt = substr($salt, 0, $required_salt_len);
 
@@ -92,9 +83,7 @@ if (!function_exists('password_hash')) {
 
 		return $ret;
 	}
-}
 
-if (!function_exists('password_get_info')) {
 	/**
 	 * Get information about the password hash. Returns an array of the information
 	 * that was used to generate the password hash.
@@ -111,7 +100,7 @@ if (!function_exists('password_get_info')) {
 	 *
 	 * @return array The array of information about the hash.
 	 */
-	function password_get_info($hash) {
+	static function get_info($hash) {
 		$return = array(
 			'algo' => 0,
 			'algoName' => 'unknown',
@@ -125,9 +114,7 @@ if (!function_exists('password_get_info')) {
 		}
 		return $return;
 	}
-}
 
-if (!function_exists('password_needs_rehash')) {
 	/**
 	 * Determine if the password hash needs to be rehashed according to the options provided
 	 *
@@ -139,8 +126,8 @@ if (!function_exists('password_needs_rehash')) {
 	 *
 	 * @return boolean True if the password needs to be rehashed.
 	 */
-	function password_needs_rehash($hash, $algo, array $options = array()) {
-		$info = password_get_info($hash);
+	static function needs_rehash($hash, $algo, array $options = array()) {
+		$info = self::get_info($hash);
 		if ($info['algo'] != $algo) {
 			return true;
 		}
@@ -154,9 +141,7 @@ if (!function_exists('password_needs_rehash')) {
 		}
 		return false;
 	}
-}
 
-if (!function_exists('password_verify')) {
 	/**
 	 * Verify a password against a hash using a timing attack resistant approach
 	 *
@@ -165,7 +150,7 @@ if (!function_exists('password_verify')) {
 	 *
 	 * @return boolean If the password matches the hash
 	 */
-    function password_verify($password, $hash) {
+	static function verify($password, $hash) {
 		if (!function_exists('crypt')) {
 			trigger_error("Crypt must be loaded for password_verify to function", E_USER_WARNING);
 			return false;
@@ -182,55 +167,52 @@ if (!function_exists('password_verify')) {
 
 		return $status === 0;
 	}
-}
-
-
-/**
- * Function to make a salt
- *
- * DO NOT USE THIS FUNCTION DIRECTLY
- *
- * @internal
- */
-function __password_make_salt($length) {
-	if ($length <= 0) {
-		trigger_error(sprintf("Length cannot be less than or equal zero: %d", $length), E_USER_WARNING);
-		return false;
-	}
-	$buffer = '';
-	$raw_length = (int) ($length * 3 / 4 + 1);
-	$buffer_valid = false;
-	if (function_exists('mcrypt_create_iv')) {
-		$buffer = mcrypt_create_iv($raw_length, MCRYPT_DEV_URANDOM);
-		if ($buffer) {
-			$buffer_valid = true;
+	
+	/**
+	 * Function to make a salt
+	 *
+	 * @internal
+	 */
+	protected static function make_salt($length) {
+		if ($length <= 0) {
+			trigger_error(sprintf("Length cannot be less than or equal zero: %d", $length), E_USER_WARNING);
+			return false;
 		}
-	}
-	if (!$buffer_valid && function_exists('openssl_random_pseudo_bytes')) {
-		$buffer = openssl_random_pseudo_bytes($raw_length);
-		if ($buffer) {
-			$buffer_valid = true;
-		}
-	}
-	if (!$buffer_valid && file_exists('/dev/urandom')) {
-		$f = @fopen('/dev/urandom', 'r');
-		if ($f) {
-			$read = strlen($buffer);
-			while ($read < $raw_length) {
-				$buffer .= fread($f, $raw_length - $read);
-				$read = strlen($buffer);
-			}
-			fclose($f);
-			if ($read >= $raw_length) {
+		$buffer = '';
+		$raw_length = (int) ($length * 3 / 4 + 1);
+		$buffer_valid = false;
+		if (function_exists('mcrypt_create_iv')) {
+			$buffer = mcrypt_create_iv($raw_length, MCRYPT_DEV_URANDOM);
+			if ($buffer) {
 				$buffer_valid = true;
 			}
 		}
-	}
-	if (!$buffer_valid) {
-		for ($i = 0; $i < $raw_length; $i++) {
-			$buffer .= chr(mt_rand(0, 255));
+		if (!$buffer_valid && function_exists('openssl_random_pseudo_bytes')) {
+			$buffer = openssl_random_pseudo_bytes($raw_length);
+			if ($buffer) {
+				$buffer_valid = true;
+			}
 		}
+		if (!$buffer_valid && file_exists('/dev/urandom')) {
+			$f = @fopen('/dev/urandom', 'r');
+			if ($f) {
+				$read = strlen($buffer);
+				while ($read < $raw_length) {
+					$buffer .= fread($f, $raw_length - $read);
+					$read = strlen($buffer);
+				}
+				fclose($f);
+				if ($read >= $raw_length) {
+					$buffer_valid = true;
+				}
+			}
+		}
+		if (!$buffer_valid) {
+			for ($i = 0; $i < $raw_length; $i++) {
+				$buffer .= chr(mt_rand(0, 255));
+			}
+		}
+		$buffer = str_replace('+', '.', base64_encode($buffer));
+		return substr($buffer, 0, $length);
 	}
-	$buffer = str_replace('+', '.', base64_encode($buffer));
-	return substr($buffer, 0, $length);
 }
